requirements.txt
# FastAPI and server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.5

# AI Providers
google-generativeai==0.7.2
openai==1.35.0

# HTTP client and utilities
httpx==0.27.0
aiohttp==3.9.5
tiktoken==0.7.0

# PDF processing
pypdf2==3.0.1

# Email service
resend==2.1.0

# Configuration and environment
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Testing (development)
pytest==8.1.1
pytest-asyncio==0.23.6
hypothesis==6.100.1